Fix custom field name translation
When expand=names is passed to the Jira REST API the response will
contain a top level entry named "names" whose value is a dict containing
the mapping between field IDs and field names.
E.g
{
  ...<Resp>,
  'names': {
    customfield_1234: "Department Goal"
  }
}

With this commit we now use this information to translate the names of custom fields in
the jira_get_issue tool before to reply to the MCP client.